+++ src/SDKs/Batch/DataPlane/Azure.Batch/GeneratedProtocol/Models/StartTask.cs
@@ -20,6 +20,19 @@ namespace Microsoft.Azure.Batch.Protocol.Models
     /// A task which is run when a compute node joins a pool in the Azure Batch
     /// service, or when the compute node is rebooted or reimaged.
     /// </summary>
+    /// <remarks>
+    /// Batch will retry tasks when a recovery operation is triggered on a
+    /// compute node. Examples of recovery operations include (but are not
+    /// limited to) when an unhealthy compute node is rebooted or a compute
+    /// node disappeared due to host failure. Retries due to recovery
+    /// operations are independent of and are not counted against the
+    /// maxTaskRetryCount. Even if the maxTaskRetryCount is 0, an internal
+    /// retry due to a recovery operation may occur. Because of this, all tasks
+    /// should be idempotent. This means tasks need to tolerate being
+    /// interrupted and restarted without causing any corruption or duplicate
+    /// data. The best practice for long running tasks is to use some form of
+    /// checkpointing.
+    /// </remarks>
     public partial class StartTask
     {
         /// <summary>
@@ -75,7 +88,11 @@ namespace Microsoft.Azure.Batch.Protocol.Models
         /// take advantage of shell features such as environment variable
         /// expansion. If you want to take advantage of such features, you
         /// should invoke the shell in the command line, for example using "cmd
-        /// /c MyCommand" in Windows or "/bin/sh -c MyCommand" in Linux.
+        /// /c MyCommand" in Windows or "/bin/sh -c MyCommand" in Linux. If the
+        /// command line refers to file paths, it should use a relative path
+        /// (relative to the task working directory), or use the Batch provided
+        /// environment variable
+        /// (https://docs.microsoft.com/en-us/azure/batch/batch-compute-node-environment-variables).
         /// </remarks>
         [JsonProperty(PropertyName = "commandLine")]
         public string CommandLine { get; set; }