+++ src/SDKs/DataFactory/Management.DataFactory/changelog.md
@@ -2,6 +2,13 @@
 
 ## Current version
 
+###  Feature Additions
+
+* Added Configure factory repository operation
+* Updated QuickBooks LinkedService to expose consumerKey and consumerSecret properties
+* Updated Several models, updating types from SecretBase to Object
+* Added Blob Events trigger
+
 ## Version 0.7.0-preview
 
 ### Feature Additions