+++ src/SDKs/DataFactory/Management.DataFactory/changelog.md
@@ -2,6 +2,17 @@
 
 ## Current version
 
+###  Feature Additions
+
+## Version 0.8.0-preview
+
+### Feature Additions
+
+* Added Configure factory repository operation
+* Updated QuickBooks LinkedService to expose consumerKey and consumerSecret properties
+* Updated Several model types from SecretBase to Object
+* Added Blob Events trigger
+
 ## Version 0.7.0-preview
 
 ### Feature Additions